iodata encoder completed
